get user repository

+++ src/main/java/com/BookMyServant/WorkerRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.BookMyServant.entity.WorkerEntity;
 
-public interface WorkerRepository extends CrudRepository<WorkerEntity, Long> {
+public interface WorkerRepository extends CrudRepository<WorkerEntity, Integer> {
